# docs/openapi.json
{
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string"
          },
          "cookieMode": {
            "type": "boolean"
          },
          "displayName": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "code",
          "displayName",
          "password"
        ],
        "type": "object"
      },
      "AuthErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "AuthTokensResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "InvitationCodeRequest": {
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "InvitationLookupResponse": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "householdName": {
            "type": "string"
          },
          "householdRole": {
            "enum": [
              "ADMIN",
              "MEMBER"
            ],
            "type": "string"
          },
          "householdsLosingProfileAccess": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "maximumAllowedRatingAge": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "mode": {
            "enum": [
              "CREATE",
              "LINK"
            ],
            "type": "string"
          },
          "profileKind": {
            "enum": [
              "KID",
              "ADULT"
            ],
            "type": "string"
          },
          "profileName": {
            "type": "string"
          },
          "profileShareOfferTargets": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "recipientEmail": {
            "type": "string"
          },
          "remainingManagers": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "RedeemPasswordResetRequest": {
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "code",
          "newPassword"
        ],
        "type": "object"
      },
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "accessCookie": {
        "description": "JWT access token. Profile tokens also satisfy SCOPE_ACCOUNT requirements. Unsafe cookie-authenticated requests also require the X-XSRF-TOKEN header.",
        "in": "cookie",
        "name": "streamarr_access",
        "type": "apiKey"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT access token. Profile tokens also satisfy SCOPE_ACCOUNT requirements.",
        "scheme": "bearer",
        "type": "http"
      },
      "playbackQuery": {
        "description": "Playback token bound to this stream session. Stream endpoints ignore bearer headers and access cookies.",
        "in": "query",
        "name": "t",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Streamarr",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/invitation/accept": {
      "post": {
        "operationId": "accept",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "invitation-controller"
        ]
      }
    },
    "/api/auth/invitation/decline": {
      "post": {
        "operationId": "decline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "invitation-controller"
        ]
      }
    },
    "/api/auth/invitation/lookup": {
      "post": {
        "operationId": "lookup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationLookupResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "invitation-controller"
        ]
      }
    },
...
    "/api/auth/password-reset/redeem": {
      "post": {
        "operationId": "redeem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemPasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "password-reset-controller"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Refusal: route on `code`, `message` is displayable"
          }
        },
        "security": [ ],
        "tags": [
          "auth-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}